remove smartos specific bits, raise up to generic solaris level
