+++ lib/chef/providers.rb
@@ -98,7 +98,6 @@ require 'chef/provider/user/pw'
 require 'chef/provider/user/useradd'
 require 'chef/provider/user/windows'
 require 'chef/provider/user/solaris'
-require 'chef/provider/user/smartos'
 require 'chef/provider/user/aix'
 
 require 'chef/provider/group/aix'
